re-arrange module with U-net architectures to improve structure
