+++ eoflow/models/__init__.py
@@ -1,3 +1,1 @@
-from .fcn_model import FCNModel
-from .tfcn_model import TFCNModel
-# from .rfcn_model import RFCNModel
+from .unets import FCNModel, TFCNModel
